Reformat TransactionEventModelTest
