+++ tests/Unit/Database/TransactionEventModelTest.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Payavel\Checkout\Tests\Unit\Database;
+
+use Payavel\Checkout\Tests\Unit\TestTransactionEventModel;
+use Payavel\Orchestration\Tests\Traits\CreatesDatabaseServiceables;
+use Payavel\Orchestration\Tests\Traits\SetsDatabaseDriver;
+
+class TransactionEventModelTest extends TestTransactionEventModel
+{
+    use CreatesDatabaseServiceables;
+    use SetsDatabaseDriver;
+}
